trivial: Fix some harmless -Wformat warnings with clang 19

+++ plugins/ch341a/fu-ch341a-cfi-device.c
@@ -57,7 +57,7 @@ fu_ch341a_cfi_device_wait_for_status_cb(FuDevice *device, gpointer user_data, GE
 			    FWUPD_ERROR_INTERNAL,
 			    "wanted 0x%x, got 0x%x",
 			    helper->value,
-			    buf[0x1] & helper->mask);
+			    (guint)(buf[0x1] & helper->mask));
 		return FALSE;
 	}
